Bug683 workaround: --toolchain=clang-usan => -O1
Clang dies on dsputil_mmx.c with -fsanitize=undefined unless it is an
optimizing compilation. Adding -O1 to the CFLAGS of --toolchain=clang-usan
works around the compilation error that otherwise occurs when Libav
is configured with --toolchain=clang-usan --disable-optimization
(the latter disables only some optimizations, but allows optimizations
which are necessary for compilation to succeed).
This workaround has been discussed with lu_zero.

http://bugzilla.libav.org/show_bug.cgi?id=683
http://llvm.org/bugs/show_bug.cgi?id=19763